pyproject.toml
[tool.poetry]
name = "sync-mastodon-follows"
version = "0.1.0"
description = "This script syncs follows between primary and secondary Mastodon accounts"
authors = ["Douglas Land <dsl@looprock.com>"]
license = "The MIT License"

[tool.poetry.dependencies]
python = "^3.10"
"Mastodon.py" = "^1.8.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
